refactor: drop outdated manifest related code
